feat(p1): compose governed frontmatter over P0
