+++ src/services/operations/obsidianNoteReplaceOperationAdapter.ts
@@ -20,6 +20,7 @@ import {
   ObsidianNoteReplaceConcurrencyError,
   ObsidianNoteReplaceJournal,
   type ObsidianNoteReplacePlan,
+  type ObsidianNoteReplaceProjection,
 } from "./obsidianNoteReplaceJournal.js";
 
 const PLAN_REF_PREFIX = "obsidian-note-replace:v1:";
@@ -74,6 +75,10 @@ export type ObsidianNoteReplacePlanInput = {
   path: string;
   nextContent: string;
   idempotencyKey: string;
+  expectedBeforeSha256?: string;
+  expectedBindingFingerprint?: string;
+  idempotencyIdentity?: string;
+  projection?: ObsidianNoteReplaceProjection;
 };
 
 function planRef(operationId: string): string {
@@ -113,6 +118,8 @@ function planDigest(plan: ObsidianNoteReplacePlan): string {
     beforeSha256: plan.beforeSha256,
     afterSha256: plan.afterSha256,
     requestDigest: plan.requestDigest,
+    idempotencyIdentity: plan.idempotencyIdentity ?? null,
+    projectionDigest: plan.projection ? operationDigest(plan.projection) : null,
   });
 }
 
@@ -187,6 +194,36 @@ function validateInput(input: ObsidianNoteReplacePlanInput): void {
   if (Buffer.byteLength(input.nextContent, "utf8") > 5 * 1024 * 1024) {
     throw new Error("nextContent exceeds the bridge limit.");
   }
+  for (const [name, value] of [
+    ["expectedBeforeSha256", input.expectedBeforeSha256],
+    ["expectedBindingFingerprint", input.expectedBindingFingerprint],
+    ["idempotencyIdentity", input.idempotencyIdentity],
+  ] as const) {
+    if (value !== undefined && !SHA256.test(value)) {
+      throw new Error(name + " must be a lowercase SHA-256 digest.");
+    }
+  }
+  if (input.projection) {
+    if (
+      input.projection.contractVersion !== 1 ||
+      !input.projection.kind ||
+      input.projection.kind.length > 128 ||
+      !input.projection.publicIdempotencyKey ||
+      input.projection.publicIdempotencyKey.length > 256 ||
+      !SHA256.test(input.projection.intentDigest) ||
+      input.projection.intentDigest !== input.idempotencyIdentity
+    ) {
+      throw new Error(
+        "projection metadata is malformed or not bound to the idempotency identity.",
+      );
+    }
+    if (
+      Buffer.byteLength(JSON.stringify(input.projection), "utf8") >
+      128 * 1024
+    ) {
+      throw new Error("projection metadata exceeds the private journal limit.");
+    }
+  }
 }
 
 export class ObsidianNoteReplaceOperationAdapter
@@ -206,10 +243,15 @@ export class ObsidianNoteReplaceOperationAdapter
       .digest("hex");
     const existing = this.journal.getByIdempotencyKey(input.idempotencyKey);
     if (existing) {
-      if (
-        existing.path !== input.path ||
-        existing.afterSha256 !== afterSha256
-      ) {
+      const identityBound =
+        existing.idempotencyIdentity !== undefined ||
+        input.idempotencyIdentity !== undefined;
+      const matches = identityBound
+        ? existing.path === input.path &&
+          existing.idempotencyIdentity !== undefined &&
+          existing.idempotencyIdentity === input.idempotencyIdentity
+        : existing.path === input.path && existing.afterSha256 === afterSha256;
+      if (!matches) {
         throw new Error(
           "The idempotency key is already bound to a different note replacement.",
         );
@@ -233,17 +275,43 @@ export class ObsidianNoteReplaceOperationAdapter
         "Atomic-write backend identity or target changed during planning.",
       );
     }
+    if (
+      input.expectedBeforeSha256 !== undefined &&
+      read.sha256 !== input.expectedBeforeSha256
+    ) {
+      throw new McpError(
+        BaseErrorCode.CONFLICT,
+        "The note changed after the domain projection was compiled.",
+      );
+    }
+    if (
+      input.expectedBindingFingerprint !== undefined &&
+      read.bindingFingerprint !== input.expectedBindingFingerprint
+    ) {
+      throw new McpError(
+        BaseErrorCode.CONFLICT,
+        "The atomic-write backend changed after the domain projection was compiled.",
+      );
+    }
     const requestDigest = operationDigest({
       operationKind: this.operationKind,
       path: input.path,
       beforeSha256: read.sha256,
       afterSha256,
       bindingFingerprint: read.bindingFingerprint,
+      idempotencyIdentity: input.idempotencyIdentity ?? null,
+      projectionDigest: input.projection
+        ? operationDigest(input.projection)
+        : null,
     });
     return receipt(
       this.journal.create({
         idempotencyKey: input.idempotencyKey,
         requestDigest,
+        ...(input.idempotencyIdentity
+          ? { idempotencyIdentity: input.idempotencyIdentity }
+          : {}),
+        ...(input.projection ? { projection: input.projection } : {}),
         path: input.path,
         beforeSha256: read.sha256,
         afterSha256,
