+++ src/services/writePolicy.ts
@@ -15,6 +15,9 @@ export type WriteOperation =
   | "obsidian_note_replace_plan"
   | "obsidian_note_replace_apply"
   | "obsidian_note_replace_recover"
+  | "obsidian_frontmatter_patch_plan"
+  | "obsidian_frontmatter_patch_apply"
+  | "obsidian_frontmatter_patch_recover"
   | "bases_create"
   | "bases_upsert_config"
   | "bases_upsert_rows"
